--- api/src/opentrons/drivers/asyncio/smoothie/__init__.py
@@ -1,5 +1,0 @@
-from .driver import SmoothieDriver
-
-__all__ = [
-    "SmoothieDriver",
-]
