+++ api/src/opentrons/hardware_control/emulation/smoothie.py
@@ -9,8 +9,7 @@ from typing import Optional, Dict
 
 from opentrons import _find_smoothie_file
 from opentrons.drivers import utils
-from opentrons.drivers.smoothie_drivers import HOMED_POSITION
-from opentrons.drivers.smoothie_drivers.driver_3_0 import GCODE
+from opentrons.drivers.smoothie_drivers.constants import GCODE, HOMED_POSITION
 from opentrons.hardware_control.emulation.parser import Command, Parser
 
 from .abstract_emulator import AbstractEmulator
