+++ api/src/opentrons/tools/args_handler.py
@@ -1,8 +1,8 @@
 import argparse
 from typing import Tuple
 
-from opentrons.hardware_control import ThreadManager, API, adapters
-from opentrons.drivers.smoothie_drivers.driver_3_0 import SmoothieDriver_3_0_0
+from opentrons.hardware_control import API
+from opentrons.drivers.smoothie_drivers import SmoothieDriver
 
 
 def root_argparser(description: str = None):
@@ -13,9 +13,9 @@ def root_argparser(description: str = None):
     return parse
 
 
-def build_driver(
+async def build_driver(
         port: str = None)\
-        -> Tuple[adapters.SynchronousAdapter, SmoothieDriver_3_0_0]:
-    hardware = ThreadManager(API.build_hardware_controller, None, port).sync
+        -> Tuple[API, SmoothieDriver]:
+    hardware = await API.build_hardware_controller(port=port)
     driver = hardware._backend._smoothie_driver
     return hardware, driver
