+++ api/tests/opentrons/drivers/asyncio/smoothie/test_parse_utils.py
@@ -1,6 +1,6 @@
 import pytest
 from opentrons.drivers import utils
-from opentrons.drivers.asyncio.smoothie import parse_utils
+from opentrons.drivers.smoothie_drivers import parse_utils
 from opentrons.drivers.utils import ParseError
