+++ api/tests/opentrons/hardware_control/integration/test_smoothie.py
@@ -6,7 +6,7 @@ from opentrons.hardware_control.emulation.app import SMOOTHIE_PORT
 from tests.opentrons.conftest import fuzzy_assert
 from opentrons.config.robot_configs import (
     DEFAULT_GANTRY_STEPS_PER_MM, DEFAULT_PIPETTE_CONFIGS, build_config)
-from opentrons.drivers.asyncio.smoothie import SmoothieDriver
+from opentrons.drivers.smoothie_drivers import SmoothieDriver
 
 
 @pytest.fixture
